REG-1690: Swagger documentation
